Added sex-specific scent tracking
Specify males and females separately (ForagerNumberFemale and ForagerNumberMale when ForagerBySex = true), females are created first. Then males and females only avoid conspecifics of their same sex, whiles are additionally attracted to females. The attraction to females is calculated as a normalized circular probablity distribution, which is added to that of the resource and then renormalized before the repulsion to males is incorporated.